Share Label State mixin functionality with IdsTextArea
