+++ src/components/ids-textarea/demos/index.yaml
@@ -8,6 +8,9 @@
   - link: example.html
     type: Main Example
     description: Showing all textarea states
+  - link: label-state.html
+    type: Example
+    description: Showing an example with hidden labels
   - link: side-by-side.html
     type: Side by Side
     description: Showing a textarea side by side working with 4.x
