+++ src/components/ids-textarea/ids-textarea-base.ts
@@ -1,4 +1,5 @@
 import IdsEventsMixin from '../../mixins/ids-events-mixin/ids-events-mixin';
+import IdsLabelStateMixin from '../../mixins/ids-label-state-mixin/ids-label-state-mixin';
 import IdsLocaleMixin from '../../mixins/ids-locale-mixin/ids-locale-mixin';
 import IdsClearableMixin from '../../mixins/ids-clearable-mixin/ids-clearable-mixin';
 import IdsDirtyTrackerMixin from '../../mixins/ids-dirty-tracker-mixin/ids-dirty-tracker-mixin';
@@ -8,11 +9,13 @@ import IdsElement from '../../core/ids-element';
 
 const Base = IdsThemeMixin(
   IdsLocaleMixin(
-    IdsValidationMixin(
-      IdsDirtyTrackerMixin(
-        IdsClearableMixin(
-          IdsEventsMixin(
-            IdsElement
+    IdsLabelStateMixin(
+      IdsValidationMixin(
+        IdsDirtyTrackerMixin(
+          IdsClearableMixin(
+            IdsEventsMixin(
+              IdsElement
+            )
           )
         )
       )
